added commentRoutes and edited post routes/home routes to better fit acceptance criteria

+++ public/js/deletePost.js
@@ -1,21 +1,23 @@
 const deletePostHandler = async (event) => {
   event.preventDefault();
-
-  const id = event.target.id;
-  console.log(`ID: ${id}`);
-
-  if (id) {
-    const response = await fetch(`/api/posts/${id}`, {
-      method: 'DELETE',
-      body: JSON.stringify({ id }),
-      headers: { 'Content-Type': 'application/json' },
-    });
-    if (response.ok) {
-      document.location.replace('/dashboard');
-    } else {
-      alert(response.statusText);
+  if (confirm("Are you sure you want to delete this post?")) {
+    const id = event.target.dataset.num;
+    console.log(`ID: ${id}`);
+  
+    if (id) {
+      const response = await fetch(`/api/posts/${id}`, {
+        method: 'DELETE',
+        body: JSON.stringify({ id }),
+        headers: { 'Content-Type': 'application/json' },
+      });
+      if (response.ok) {
+        document.location.replace('/dashboard');
+      } else {
+        alert(response.statusText);
+      }
     }
   }
-}
+  };
+
 
 document.querySelector('#delete-btn').addEventListener('click', deletePostHandler);
